Add `cordy-web` subproject for compiling a REPL to be used with web assembly. Several other changes and fixes are included:
- Refactor the `repl` module to be not entirely callback based, but instead based on a `Repl` struct that can be invoked sequentially.
- Expose `scanner::scan` to be able to do syntax highlighting of source code without requiring a full compile.
- Add proper error handling for `read_text` and `write_text` instead of a panic,
- Fix issues with `ValuePtr` representations using 32-bit and 64-bit types interchangeably, and fix issues with `RuntimeError` not being fully integrated as a owned type.

+++ cordy-sys/src/vm/mod.rs
@@ -287,7 +287,7 @@ impl<R, W> VirtualMachine<R, W> where
                 self.push(self.stack[local].clone());
             }
             StoreLocal(local, pop) => {
-                trace::trace_interpreter!("vm::run StoreLocal index={}, value={}, prev={}", local, self.stack.last().unwrap().as_debug_str(), self.stack[local].as_debug_str());
+                trace::trace_interpreter!("vm::run StoreLocal index={}, value={}, prev={}", local, self.stack.last().unwrap().as_debug_str(), self.stack[local as usize].as_debug_str());
                 let value = if pop { self.pop() } else { self.peek(0).clone() };
                 self.store_local(local, value);
             },
@@ -302,7 +302,7 @@ impl<R, W> VirtualMachine<R, W> where
                 }
             },
             StoreGlobal(local, pop) => {
-                trace::trace_interpreter!("vm::run StoreGlobal index={}, value={}, prev={}", local, self.stack.last().unwrap().as_debug_str(), self.stack[local].as_debug_str());
+                trace::trace_interpreter!("vm::run StoreGlobal index={}, value={}, prev={}", local, self.stack.last().unwrap().as_debug_str(), self.stack[local as usize].as_debug_str());
                 let value = if pop { self.pop() } else { self.peek(0).clone() };
                 self.store_global(local, value)?;
             },
